project09/requirements.txt
flask>=2.0.0
tensorflow==2.6.2
numpy==1.19.5
pillow>=8.0.0,<10.0.0
flask-cors>=3.0.10
protobuf>=3.9.2,<4.0.0
keras==2.6.0

